requirements.txt
pytest
mypy
flake8
antlr4-tools
antlr4-python3-runtime